requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain==0.1.6
langchain-community==0.0.19
langchain-core==0.1.21
langgraph==0.0.20
langchain-anthropic==0.1.1
langchain-google-genai==1.0.1

# Vector Store
pinecone-client==3.0.2
langchain-pinecone==0.0.1

# Embeddings
sentence-transformers==2.3.1

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Redis for session/cache
redis==5.0.1
hiredis==2.3.2

# Environment & Config
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tenacity==8.2.3

# Monitoring & Logging
loguru==0.7.2

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
